ingestion_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
requests>=2.31.0
cloudinary>=1.36.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
redis>=5.0.0
